add comments and fix sample methods name
