+++ x/launch/keeper/request.go
@@ -1,10 +1,13 @@
 package keeper
 
 import (
+	"fmt"
 	"strconv"
 
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	spnerrors "github.com/tendermint/spn/pkg/errors"
 	"github.com/tendermint/spn/x/launch/types"
 )
 
@@ -111,4 +114,78 @@ func (k Keeper) GetAllRequest(ctx sdk.Context) (list []types.Request) {
 	}
 
 	return
+}
+
+// applyRequest approves the request and performs
+// the launch information changes
+func applyRequest(
+	ctx sdk.Context,
+	k Keeper,
+	msg *types.MsgSettleRequest,
+	request types.Request,
+) error {
+	cdc := codectypes.NewInterfaceRegistry()
+	types.RegisterInterfaces(cdc)
+
+	var content types.RequestContent
+	if err := cdc.UnpackAny(request.Content, &content); err != nil {
+		return spnerrors.Critical(err.Error())
+	}
+
+	switch c := content.(type) {
+	case *types.GenesisAccount:
+		_, foundGenesis := k.GetGenesisAccount(ctx, msg.ChainID, c.Address)
+		_, foundVested := k.GetVestedAccount(ctx, msg.ChainID, c.Address)
+		if foundVested || foundGenesis {
+			return spnerrors.Critical(fmt.Sprintf(
+				"account %s for chain %s already exist",
+				c.Address, msg.ChainID),
+			)
+		}
+		k.SetGenesisAccount(ctx, *c)
+	case *types.VestedAccount:
+		_, foundGenesis := k.GetGenesisAccount(ctx, msg.ChainID, c.Address)
+		_, foundVested := k.GetVestedAccount(ctx, msg.ChainID, c.Address)
+		if foundVested || foundGenesis {
+			return spnerrors.Critical(fmt.Sprintf(
+				"account %s for chain %s already exist",
+				c.Address, msg.ChainID),
+			)
+		}
+		k.SetVestedAccount(ctx, *c)
+	case *types.AccountRemoval:
+		_, foundGenesis := k.GetGenesisAccount(ctx, msg.ChainID, c.Address)
+		_, foundVested := k.GetVestedAccount(ctx, msg.ChainID, c.Address)
+		if foundGenesis {
+			k.RemoveGenesisAccount(ctx, msg.ChainID, c.Address)
+		}
+		if foundVested {
+			k.RemoveVestedAccount(ctx, msg.ChainID, c.Address)
+		}
+		if !foundGenesis && !foundVested {
+			return spnerrors.Critical(fmt.Sprintf(
+				"account %s for chain %s not found",
+				c.Address, msg.ChainID),
+			)
+		}
+	case *types.GenesisValidator:
+		if _, found := k.GetGenesisValidator(ctx, msg.ChainID, c.Address); found {
+			return spnerrors.Critical(fmt.Sprintf(
+				"genesis validator %s for chain %s already exist",
+				c.Address, msg.ChainID),
+			)
+		}
+		k.SetGenesisValidator(ctx, *c)
+	case *types.ValidatorRemoval:
+		if _, found := k.GetGenesisValidator(ctx, msg.ChainID, c.ValAddress); !found {
+			return spnerrors.Critical(fmt.Sprintf(
+				"genesis validator %s for chain %s not found",
+				c.ValAddress, msg.ChainID),
+			)
+		}
+		k.RemoveGenesisValidator(ctx, msg.ChainID, c.ValAddress)
+	default:
+		return spnerrors.Critical("unknown request content type")
+	}
+	return nil
 }
